requirements.txt
# Root level requirements for shared utilities
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.18.1
google-cloud-pubsub==2.18.4
google-cloud-monitoring==2.17.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
pandas==2.1.4
numpy==1.24.4
python-dateutil==2.8.2
pytz==2023.3
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.25.2